cube

+++ cube.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cube.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: houaslam <houaslam@student.42.fr>          +#+  +:+       +#+        */
+/*   By: fadermou <fadermou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/19 11:29:31 by houaslam          #+#    #+#             */
-/*   Updated: 2023/10/01 15:44:51 by houaslam         ###   ########.fr       */
+/*   Updated: 2023/10/01 18:27:30 by fadermou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -146,8 +146,8 @@ typedef struct s_directions
 	char			*so;
 	char			*we;
 	char			*ea;
-	int				f;
-	int				c;
+	long long		f;
+	long long		c;
 	struct s_map	*map;
 }	t_directions;
